[FEATURE] Introduce nested node registry
In order to keep resolvers and arguments and types close together, nested nodes were implemented.

A nested node, for instance, nested-language-node, can be easily reused by any model/record which needs to nest a language node.

In order to resolve the issue of circular-references (nested-page-node needs page-type but page-type also needs nested-page-node) a nested-node-registry was introduced, which loads nested nodes on runtime.