WIP: Full type safe flexform integration
Playing around with a full-type-supported flex-form integration.

Flexform by nature is not type safe and unstructured. This means even the same field depending on ds_pointerFields can have complete different types. This can be solved by nesting the different ds defintion or graphql interfaces. However, to be honest, i am not sure if the effort is worth it, as it brings maybe unnecessary complexity to the code as anybody could also extend his record and easily integrate needed flexform fields for his context.