[TASK] Explicitly disallow usage of dynamic flex-form definitions
In TYPO3 it is possible to skip the ds configuration for flex-form fields. This means that the flex form definition is directly stored on a record.

This kind of flex-form definitions cannot be supported and have to be self-implemented.

A test for specifying this type of flex-form definition as "unsupported" is implemented here.