[FIX] Fixed content rendering returning empty string on TYPO3 v12.1.*
Content rendering on TYPO3 is not easy on runtime. Imagine to content of two different pages with complete different typo-script configuration on each page.

Furthermore, with each version of TYPO3 the api of TSFE is changing. Therefore, I split up the content-render-record type in multiple methods dedicated to a specific typo3 version.

By this compatibility for 12.1 is also provided, where a couple of methods of TYPO3 have changed:

- RootlineUtility::purgeCaches was removed
- $tsfe->getFromCache now returns a request with typo-script object as an attribute, which needs to be passed to content object renderer.

Additional:

Also fixed minor bugs on test-app site configuration files.