[BUGFIX] Fix content-renderer compatibility to Typo3 version > 12.1
Newer versions like 12.2 would not be catched in ContentRenderRecordTypeBuilderExtender.php as it was only checking version 12.1 and there was a bug on Typo3Environment::isVersion method. This resulted into an error complaining about missing purgeCaches which should never be called above 12.1.

---

Further: php-cs-fixer is no longer part of main package, it is instead install on-the-fly within tools directory.

This way we avoid composer conflicts to effect the application, which are only related to php-cs-fixer.

Also phpunit is now updated to version 10